requirements.txt
discord.py
feedparser
requests
python-dotenv